pridany operatory a testy

+++ tests/test_limits.h
@@ -17,16 +17,16 @@ inline double calc_min(uint8_t numBits, uint8_t denBits)
     return -exp2(numBits + denBits - 1) / exp2(denBits);
 }
 
-#define CHECK_MIN(n, d)                                 \
-    do {                                                \
-        q<n, d> f;                                      \
-        EXPECT_NEAR(f.min(), calc_min(n, d), exp2(-d)); \
+#define CHECK_MIN(n, d)                                            \
+    do {                                                           \
+        q<n, d> f;                                                 \
+        EXPECT_NEAR(f.min().toDouble(), calc_min(n, d), exp2(-d)); \
     } while (0);
 
-#define CHECK_MAX(n, d)                                 \
-    do {                                                \
-        q<n, d> f;                                      \
-        EXPECT_NEAR(f.max(), calc_max(n, d), exp2(-d)); \
+#define CHECK_MAX(n, d)                                            \
+    do {                                                           \
+        q<n, d> f;                                                 \
+        EXPECT_NEAR(f.max().toDouble(), calc_max(n, d), exp2(-d)); \
     } while (0);
 
 class limits_min8;
@@ -38,6 +38,5 @@ class limits_max8;
 class limits_max16;
 class limits_max32;
 class limits_max64;
-
 
 #endif // TEST_LIMITS_H
